Support 64-bit integers across Lua-Java boundaries

+++ example/src/test/java/party/iroiro/luajava/JuaApiTest.java
@@ -70,7 +70,7 @@ public class JuaApiTest {
 
     private static class NativeTest extends Lua51Natives {
         @Override
-        public long lua_newuserdata(long ptr, int size) {
+        public long lua_newuserdata(long ptr, long size) {
             return super.lua_newuserdata(ptr, size);
         }
     }
